+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.13.0](https://github.com/allenai/beaker-py/releases/tag/v0.13.0) - 2022-04-27
+
 ### Added
 
 - Added `Beaker.dataset.url()` method.